refactor: changing the path archieves
