+++ packages/ui/src/ui/Typegraph/Heading/index.tsx
@@ -1,7 +1,7 @@
 import { ElementType, ReactNode } from 'react'
 import { styles } from './styles'
 import { VariantProps } from 'tailwind-variants'
-import '../../styles/global.css'
+import '../../../styles/global.css'
 
 export interface HeadingProps extends VariantProps<typeof styles> {
   children: ReactNode
