+++ packages/ui/src/ui/Typegraph/Heading/styles.ts
