+++ packages/ui/src/ui/Typegraph/Text/index.tsx
@@ -1,7 +1,7 @@
 import { ElementType, ReactNode } from 'react'
 import { styles } from './styles'
 import { VariantProps } from 'tailwind-variants'
-import '../../styles/global.css'
+import '../../../styles/global.css'
 
 export interface TextProps extends VariantProps<typeof styles> {
   children: ReactNode
