+++ packages/ui/src/ui/Typegraph/Text/styles.ts
